fix minor bugs + migrate to pug
