--- .github/workflows/keep-alive.yml
@@ -1,14 +1,0 @@
-name: Keep Alive Workflow
-
-on:
-    schedule:
-        - cron: "0 0 1 */1 *" #once a month
-
-jobs:
-    keep-alive:
-        runs-on: ubuntu-latest
-        permissions:
-            actions: write
-        steps:
-            - uses: actions/checkout@v4
-            - uses: gautamkrishnar/keepalive-workflow@2.0.7
