+++ hours-email-reminder/email.pug
@@ -1,0 +1,70 @@
+doctype html
+html(lang="en" dir="ltr" xmlns="http://www.w3.org/1999/xhtml" xmlns:o="urn:schemas-microsoft-com:office:office")
+  head
+    meta(charset="UTF-8")
+    meta(name="viewport" content="width=device-width, initial-scale=1")
+    meta(name="x-apple-disable-message-reformatting")
+    meta(http-equiv="X-UA-Compatible" content="IE=edge")
+    meta(name="format-detection" content="telephone=no")
+    title Copy of (2) New Message
+    style.
+      .rollover:hover .rollover-first { max-height:0px!important; display:none!important; }
+      .rollover:hover .rollover-second { max-height:none!important; display:block!important; }
+      .rollover span { font-size:0px; }
+      u + .body img ~ div div { display:none; }
+      #outlook a { padding:0; }
+      span.MsoHyperlink, span.MsoHyperlinkFollowed { color:inherit; mso-style-priority:99; }
+      a.es-button { mso-style-priority:100!important; text-decoration:none!important; }
+      a[x-apple-data-detectors], #MessageViewBody a { color:inherit!important; text-decoration:none!important; font-size:inherit!important; font-family:inherit!important; font-weight:inherit!important; line-height:inherit!important; }
+      .es-desk-hidden { display:none; float:left; overflow:hidden; width:0; max-height:0; line-height:0; mso-hide:all; }
+  body(style="width:100%;height:100%;-webkit-text-size-adjust:100%;-ms-text-size-adjust:100%;padding:0;margin:0")
+    .es-wrapper-color(style="background-color:#F6F6F6")
+      table.es-wrapper(role="none" cellspacing="0" cellpadding="0" width="100%" style="border-collapse:collapse;width:100%;height:100%;background-repeat:repeat;background-position:center top;background-color:#F6F6F6")
+        tr
+          td(valign="top")
+            table.es-header(role="none" cellspacing="0" cellpadding="0" align="center" style="width:100%;background-color:transparent")
+              tr
+                td(align="center")
+                  table.es-header-body(role="none" cellspacing="0" cellpadding="0" align="center" bgcolor="#ffffff" style="background-color:#FFFFFF;width:600px")
+                    tr
+                      td(align="left" style="padding:20px 150px")
+                        table.es-right(role="none" cellspacing="0" cellpadding="0" align="right" style="float:right")
+                          tr
+                            td(align="left" style="width:300px")
+                              table(role="presentation" width="100%" cellspacing="0" cellpadding="0")
+                                tr
+                                  td(align="center" style="font-size:0")
+                                    img.adapt-img(src="https://epiujlr.stripocdn.email/content/guids/CABINET_b671897a35158e25b88e73b31b95f12e48ace5edc8e5b734899d083c66476053/images/h4hlogo_zhD.png" alt="" width="150" height="161" style="display:block;border:0")
+
+            table.es-content(role="none" cellspacing="0" cellpadding="0" align="center" style="width:100%")
+              tr
+                td(align="center")
+                  table.es-content-body(role="none" cellspacing="0" cellpadding="0" align="center" bgcolor="#ffffff" style="background-color:#FFFFFF;width:600px")
+                    tr
+                      td(align="left" style="padding:20px")
+                        table(role="none" width="100%" cellspacing="0" cellpadding="0")
+                          tr
+                            td(align="center" style="width:560px")
+                              table(role="presentation" width="100%" cellspacing="0" cellpadding="0")
+                                tr
+                                  td(align="left")
+                                    p This is a friendly reminder that completing 800 hours of training is required in order to attend our 4-day advanced training at Camp Grace in Mobile, Alabama! Please make sure to track your progress and work towards the 800-hour goal. This training is essential to help you gain the skills and knowledge needed for success in the program. Thank you for your commitment, and we look forward to seeing you at Camp Grace!
+
+            table.es-footer(role="none" cellspacing="0" cellpadding="0" align="center" style="width:100%;background-color:transparent")
+              tr
+                td(align="center")
+                  table.es-footer-body(role="none" cellspacing="0" cellpadding="0" align="center" bgcolor="#ffffff" style="background-color:#FFFFFF;width:600px")
+                    tr
+                      td(align="left" style="padding:20px")
+                        table.es-left(role="none" cellspacing="0" cellpadding="0" align="left" style="float:left")
+                          tr
+                            td(align="left" class="es-m-p20b" style="width:270px")
+                              table(role="none" width="100%" cellspacing="0" cellpadding="0")
+                                tr
+                                  td(align="center" style="display:none")
+                        table.es-right(role="none" cellspacing="0" cellpadding="0" align="right" style="float:right")
+                          tr
+                            td(align="left" style="width:270px")
+                              table(role="none" width="100%" cellspacing="0" cellpadding="0")
+                                tr
+                                  td(align="center" style="display:none")
