+++ hours-email-reminder/package.json
@@ -3,7 +3,7 @@
   "version": "1.0.0",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "start": "npx tsx send-hour-email.ts"
   },
   "keywords": [],
   "author": "",
@@ -11,6 +11,12 @@
   "description": "",
   "dependencies": {
     "juno-sdk": "^0.0.5",
-    "mongoose": "^8.8.1"
+    "mongoose": "^8.8.1",
+    "pug": "^3.0.3"
+  },
+  "devDependencies": {
+    "@types/node": "^22.9.0",
+    "@types/pug": "^2.0.10",
+    "tsx": "^4.19.1"
   }
 }
